[feature/#24] core:domain:chat 모듈 생성
