--- chat/src/androidMain/kotlin/com/nexters/emotia/core/data/chat/Platform.android.kt
@@ -1,3 +1,0 @@
-package com.nexters.emotia.core.data.chat
-
-actual fun platform() = "Android"
