--- chat/src/commonMain/kotlin/com/nexters/emotia/core/data/chat/Platform.kt
@@ -1,3 +1,0 @@
-package com.nexters.emotia.core.data.chat
-
-expect fun platform(): String
