--- chat/src/iosMain/kotlin/com/nexters/emotia/core/data/chat/Platform.ios.kt
@@ -1,3 +1,0 @@
-package com.nexters.emotia.core.data.chat
-
-actual fun platform() = "iOS"
