+++ core/data/chat/src/commonMain/kotlin/com/nexters/emotia/core/data/chat/repository/ChattingRepositoryImpl.kt
@@ -2,9 +2,9 @@ package com.nexters.emotia.core.data.chat.repository
 
 import com.nexters.emotia.core.data.chat.datasource.ChattingRemoteDataSource
 import com.nexters.emotia.core.data.chat.mapper.toDomain
-import com.nexters.emotia.domain.chat.ChattingRepository
-import com.nexters.emotia.domain.chat.entity.ChatMessage
-import com.nexters.emotia.domain.chat.entity.ChattingRoom
+import com.nexters.emotia.core.domain.chat.entity.ChatMessage
+import com.nexters.emotia.core.domain.chat.entity.ChattingRoom
+import com.nexters.emotia.core.domain.chat.repsitory.ChattingRepository
 
 class ChattingRepositoryImpl(
     private val remoteDataSource: ChattingRemoteDataSource
