+++ core/domain/chat/.gitignore
