+++ core/domain/chat/src/androidMain/AndroidManifest.xml
