+++ core/domain/chat/src/androidMain/kotlin/com/nexters/emotia/core/domain/chat/Platform.android.kt
@@ -1,0 +1,5 @@
+package com.nexters.emotia.core.domain.chat
+
+actual fun platform(): String {
+    TODO("Not yet implemented")
+}
