+++ core/domain/chat/src/commonMain/kotlin/com/nexters/emotia/core/domain/chat/Platform.kt
@@ -1,0 +1,3 @@
+package com.nexters.emotia.core.domain.chat
+
+expect fun platform(): String
