+++ core/domain/chat/src/commonMain/kotlin/com/nexters/emotia/core/domain/chat/entity/ChatMessage.kt
@@ -1,0 +1,5 @@
+package com.nexters.emotia.core.domain.chat.entity
+
+data class ChatMessage(
+    val message: String
+)
