+++ core/domain/chat/src/commonMain/kotlin/com/nexters/emotia/core/domain/chat/entity/ChattingRoom.kt
@@ -1,0 +1,6 @@
+package com.nexters.emotia.core.domain.chat.entity
+
+data class ChattingRoom(
+    val roomId: Int,
+    val firstMessage: String
+)
