+++ core/domain/chat/src/iosMain/kotlin/com/nexters/emotia/core/domain/chat/Platform.ios.kt
@@ -1,0 +1,3 @@
+package com.nexters.emotia.core.domain.chat
+
+actual fun platform() = "iOS"
