--- core/domain/src/commonMain/kotlin/com/nexters/emotia/domain/Platform.kt
@@ -1,3 +1,0 @@
-package com.nexters.emotia.domain
-
-// expect fun platform(): String
