--- core/domain/src/commonMain/kotlin/com/nexters/emotia/domain/chat/ChattingRepository.kt
@@ -1,9 +1,0 @@
-package com.nexters.emotia.domain.chat
-
-import com.nexters.emotia.domain.chat.entity.ChatMessage
-import com.nexters.emotia.domain.chat.entity.ChattingRoom
-
-interface ChattingRepository {
-    suspend fun createRoom(username: String): Result<ChattingRoom>
-    suspend fun sendChat(roomId: Int, message: String): Result<ChatMessage>
-}
