--- core/domain/src/commonMain/kotlin/com/nexters/emotia/domain/chat/entity/ChatMessage.kt
@@ -1,5 +1,0 @@
-package com.nexters.emotia.domain.chat.entity
-
-data class ChatMessage(
-    val message: String
-)
