+++ feature/main/build.gradle.kts
@@ -8,8 +8,9 @@ kotlin {
             implementation(projects.feature.onboarding)
             implementation(projects.feature.chatting)
             implementation(projects.feature.result)
-            implementation(projects.core.domain)
+
             implementation(projects.core.designsystem)
+
             implementation(libs.koin.core)
             implementation(libs.koin.compose.viewmodel)
             implementation(libs.slf4j.api)
