+++ settings.gradle.kts
@@ -33,9 +33,11 @@ include(":composeApp")
 
 include(":core:designsystem")
 include(":core:model:fairybook")
+
 include(":core:data:chat")
 
-include(":core:domain")
+include(":core:domain:chat")
+
 include(":core:network")
 include(":core:navigation")
